+++ crates/devela/src/text/unicode/grapheme/u8.rs
@@ -11,7 +11,7 @@ use crate::{GraphemeMachine, GraphemeScanner, StringU8};
 /* definitions */
 
 #[doc = crate::_tags!(text)]
-#[doc = concat!["An ", crate::_ABBR_EGC!(), " backed by a [`StringU8`]."]]
+/// An extended grapheme cluster backed by a [`StringU8`].
 #[doc = crate::_doc_meta!{
     location("text/unicode/grapheme", struct GraphemeU8),
     test_size_of(GraphemeU8<15> = 16|128; niche Option),